pyproject.toml
[project]
name = "rubaiyat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.22",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]
